Email trigger: Allow Follow-Up messages
